fix: cleaned splash screen
